up-g-5

+++ python/完全数チェッカー.py
@@ -1,6 +1,6 @@
 def pn(num):
     num_keep,li=num,[]
-    for i in range(1,(num**(1/2))+1):
+    for i in range(1,int(num**(1/2))+1):
         if num%i==0:
             li.append(i)
         if i!=num//i:
